# Conflicts:
#	.gitignore
#	revoke_debt.sql

+++ VS_LOAN.Core.Web/Controllers/MCreditController.cs
@@ -177,7 +177,7 @@ namespace VS_LOAN.Core.Web.Controllers
                         HosoId = id,
                         Noidung = model.LastNote,
                         CommentTime = DateTime.Now,
-                        TypeId = NoteType.MCreditTemp
+                        TypeId = (int)NoteType.MCreditTemp
                     };
                    await _rpNote.AddNoteAsync(note);
                 }
@@ -215,7 +215,7 @@ namespace VS_LOAN.Core.Web.Controllers
                     HosoId = model.Id,
                     Noidung = model.LastNote,
                     CommentTime = DateTime.Now,
-                    TypeId = NoteType.MCreditTemp
+                    TypeId = (int)NoteType.MCreditTemp
                 };
                 await _rpNote.AddNoteAsync(note);
             }
